+++ server/camcops_server/tasks/perinatalpoem.py
@@ -27,10 +27,9 @@ camcops_server/tasks/perinatalpoem.py
 """
 
 import re
-from typing import Dict, Generator, List, Tuple, Type
+from typing import Dict, List, Tuple, Type
 
 from cardinal_pythonlib.classes import classproperty
-import pendulum
 from pyramid.renderers import render_to_response
 from pyramid.response import Response
 from sqlalchemy.sql.expression import and_, column, select
@@ -63,7 +62,6 @@ from camcops_server.cc_modules.cc_task import (
 )
 from camcops_server.cc_modules.cc_text import SS
 from camcops_server.cc_modules.cc_tsv import TsvPage
-from camcops_server.cc_modules.cc_unittest import DemoDatabaseTestCase
 
 
 # =============================================================================
@@ -722,139 +720,3 @@ class PerinatalPoemReport(DateTimeFilteredReportMixin, Report,
         A list of all the additional comments.
         """
         return [x[0] for x in self._get_comment_rows(req)]
-
-
-# =============================================================================
-# Unit tests
-# =============================================================================
-
-class PerinatalPoemReportTestCase(DemoDatabaseTestCase):
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.id_sequence = self.get_id()
-
-    def setUp(self) -> None:
-        super().setUp()
-
-        self.report = PerinatalPoemReport()
-
-        # Really only needed for tests
-        self.report.start_datetime = None
-        self.report.end_datetime = None
-
-    @staticmethod
-    def get_id() -> Generator[int, None, None]:
-        i = 1
-
-        while True:
-            yield i
-            i += 1
-
-    def create_task(self, **kwargs) -> None:
-        task = PerinatalPoem()
-        self.apply_standard_task_fields(task)
-        task.id = next(self.id_sequence)
-
-        era = kwargs.pop('era', None)
-        if era is not None:
-            task.when_created = pendulum.parse(era)
-
-        for name, value in kwargs.items():
-            setattr(task, name, value)
-
-        self.dbsession.add(task)
-
-
-class PerinatalPoemReportTests(PerinatalPoemReportTestCase):
-    """
-    Most of the base class tested in APEQCPFT Perinatal so just some basic
-    sanity checking here
-    """
-
-    def create_tasks(self):
-        self.create_task(general_comments="comment 1")
-        self.create_task(general_comments="comment 2")
-        self.create_task(general_comments="comment 3")
-
-        self.dbsession.commit()
-
-    def test_qa_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[0].rows
-
-        self.assertEqual(len(rows), 1)
-        self.assertEqual(len(rows[0]), 4)
-
-    def test_qb_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[1].rows
-
-        self.assertEqual(len(rows), 1)
-        self.assertEqual(len(rows[0]), 4)
-
-    def test_q1_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[2].rows
-
-        self.assertEqual(len(rows), 2)
-        self.assertEqual(len(rows[0]), 7)
-
-    def test_q2_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[3].rows
-
-        self.assertEqual(len(rows), 12)
-        self.assertEqual(len(rows[0]), 6)
-
-    def test_q3_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[4].rows
-
-        self.assertEqual(len(rows), 6)
-        self.assertEqual(len(rows[0]), 6)
-
-    def test_participation_rows_counts(self) -> None:
-        tables = self.report._get_html_tables(self.req)
-
-        rows = tables[5].rows
-
-        self.assertEqual(len(rows), 1)
-        self.assertEqual(len(rows[0]), 4)
-
-    def test_comments(self) -> None:
-        expected_comments = [
-            "comment 1", "comment 2", "comment 3",
-        ]
-        comments = self.report._get_comments(self.req)
-        self.assertEqual(comments, expected_comments)
-
-
-class PerinatalPoemReportDateRangeTests(PerinatalPoemReportTestCase):
-    def create_tasks(self) -> None:
-        self.create_task(general_comments="comments 1",
-                         era="2018-10-01T00:00:00.000000+00:00")
-        self.create_task(general_comments="comments 2",
-                         era="2018-10-02T00:00:00.000000+00:00")
-        self.create_task(general_comments="comments 3",
-                         era="2018-10-03T00:00:00.000000+00:00")
-        self.create_task(general_comments="comments 4",
-                         era="2018-10-04T00:00:00.000000+00:00")
-        self.create_task(general_comments="comments 5",
-                         era="2018-10-05T00:00:00.000000+00:00")
-        self.dbsession.commit()
-
-    def test_comments_filtered_by_date(self) -> None:
-        self.report.start_datetime = "2018-10-02T00:00:00.000000+00:00"
-        self.report.end_datetime = "2018-10-05T00:00:00.000000+00:00"
-
-        comments = self.report._get_comments(self.req)
-        self.assertEqual(len(comments), 3)
-
-        self.assertEqual(comments[0], "comments 2")
-        self.assertEqual(comments[1], "comments 3")
-        self.assertEqual(comments[2], "comments 4")
